fix individ_tg_bot
