+++ settings/config.ini
@@ -4,6 +4,8 @@ user=postgres
 password=00000
 host=127.0.0.1
 port=5432
+DB_URL=postgresql://postgres:00000@127.0.0.1:5432/postgres
+;DB_URL=dialect+driver://username:password@host:port/database
 
 [My_channels]
 no_sorted_channel = -1001175723940
